закончил с подписками

+++ backend/api/views.py
@@ -49,43 +49,43 @@ class FollowViewSet(UserViewSet):
     pagination_class = PageNumberPagination
 
     @action(
-        methods=['post'], detail=True, permission_classes=[IsAuthenticated])
+        methods=['post', 'delete'],
+        detail=True,
+        permission_classes=[IsAuthenticated])
+    
     def subscribe(self, request, id=None):
         user = request.user
         author = get_object_or_404(User, id=id)
+        if request.method == 'POST':
+            if user == author:
+                return Response({
+                    'errors': 'Нельзя подписываться на себя'
+                }, status=status.HTTP_400_BAD_REQUEST)
+            if Follow.objects.filter(user=user, author=author).exists():
+                return Response({
+                    'errors': 'Вы уже подписаны на пользователя'
+                }, status=status.HTTP_400_BAD_REQUEST)
 
-        if user == author:
-            return Response({
-                'errors': 'Ошибка подписки, нельзя подписываться на себя'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        if Follow.objects.filter(user=user, author=author).exists():
-            return Response({
-                'errors': 'Ошибка подписки, вы уже подписаны на пользователя'
-            }, status=status.HTTP_400_BAD_REQUEST)
-
-        follow = Follow.objects.create(user=user, author=author)
-        serializer = FollowSerializer(
-            follow, context={'request': request}
-        )
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    @subscribe.mapping.delete
-    def del_subscribe(self, request, id=None):
-        user = request.user
-        author = get_object_or_404(User, id=id)
-        if user == author:
-            return Response({
-                'errors': 'Ошибка отписки, нельзя отписываться от самого себя'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        follow = Follow.objects.filter(user=user, author=author)
-        if not follow.exists():
-            return Response({
-                'errors': 'Ошибка отписки, вы уже отписались'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        follow.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            follow = Follow.objects.create(user=user, author=author)
+            serializer = FollowSerializer(
+                follow, context={'request': request}
+            )
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.method == 'DELETE':
+            if user == author:
+                return Response({
+                    'errors': 'Нельзя отписываться от самого себя'
+                }, status=status.HTTP_400_BAD_REQUEST)
+            follow = Follow.objects.filter(user=user, author=author)
+            if not follow.exists():
+                return Response({
+                    'errors': 'Вы уже отписались'
+                }, status=status.HTTP_400_BAD_REQUEST)
+            follow.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(detail=False, permission_classes=[IsAuthenticated])
+    
     def subscriptions(self, request):
         user = request.user
         queryset = Follow.objects.filter(user=user)
